Specify response status to authenticated/unauthenticated requests
A request that is unauthenticated (no authorization header, an empty authorization header, or an authorization header with an invalid signature) will result in an HTTP status 404 response. A request that is authenticated with a valid digital signature (according to the scheme in README.md) will result in an HTTP status 200 response.