Delegate authentication to single-responsibility entities
This commit refactors authentication logic away from the SecretMessagesController so that it is handled by more modular entities.

Verification of the digital signature is handled by an Authentication module. Filtering out unauthenticated HTTP requests is handled by an AuthenticationController. The only responsibility of the SecretMessagesController is to get resources and return them in an HTTP response.