Add unit specs for the Authentication module
This commit adds unit specs for the #authenticated?(request) method of the Authentication module. In order for this method to return true, the request object must have properties of an HTTP method and a path that, when hashed with a key, will produce a hash that matches the value of the request's Authorization header.

This commit also makes a couple of minor edits to comments:
  * Refer to the key as a key instead of a password
  * Show the HTTP version in the request line as HTTP/1.1 instead of just 1.1