Specify a module for encrypting data
Specify a module called Encryption with a method #encrypt(data, iv). The first argument is a string representing data to encrypt. The second argument is an optional string representing the AES initialization vector. This method returns a Ruby Hash with a key :cipher and corresponding value of encrypted data, and a key :iv and corresponding value of the initialization vector.