Improve testability of an encrypted HTTP response
The prior signature of Encryption#encrypt(data, initialization_vector) would have made it difficult or impossible to test the contents of an encrypted HTTP response in the request specs. This is because #encrypt would not be called in the request spec, meaning the initialization vector could not be set in that environment, meaning a random initialization vector would be used each time and the encrypted HTTP response would be different each time. Although this is the desired production behavior, it makes it difficult to test the exact contents of an encrypted HTTP response.

This commit applies changes that remove initialization_vector from the arguments of #encrypt. Instead, the initialization vector comes from an environment variable if that environment variable exists, and otherwise it is randomly generated within #encrypt. This means that both the encryption key and the initialization vector can be set in the test environment via environment variables, which will result in a deterministic and testable encrypted HTTP response. In production, there will be no environment variable for the initialization vector and it will instead be randomly generated for each response, which is preferable as per the Ruby OpenSSL docs.