Add a spec for requesting all secret message resources
When an authenticated request is made for all secret message resources, the response body should be a JSON string where the top-level key is the resource's id, and the corresponding value is a JSON object with key-value pairs for the encrypted message and the encryption's AES initialization vector.

Previously, in the setup for requesting a single resource, the initialization vector was passed in the HTTP response via a header. That architecture will not work with a request for several resources, because a different initialization vector is used to encrypt each resource's message. So, for the sake of consistency across the application, a reasonable change to make soon will be to pass the initialization vector within the JSON object in the response for all requests (including requests for both single and multiple resources).