Send the AES initialization vector in the response body
This commit simplifies some controller code and introduces more consistency across the application.

Previously, the AES initialization vector was sent in a response header, which would not work for a request for multiple resources that are encrypted using a unique initialization vector for each encryption. This commit applies changes that send the initialization vector inside the JSON of the response body, so that the initialization vector for each encrypted message can be associated with that encryption. This architecture allows the exact output of Encryption#encrypt to be sent as JSON, so another small change in this commit is to change the output of Encryption#encrypt to use a key called :message instead of :cipher, which conforms to the existing API and simplifies the controller code for a single-resource request.