Specify that the JSON response should contain an array of resources
This commit enables the client to more easily parse the JSON response to requests for multiple and single resources.

Previously, the response to a request for a single resource would be a JSON object with no nesting, and the response to a request for multiple resources would be a JSON object containing multiple nested objects. This dual structure would have required the client to use a different set of rules to parse the JSON for a single resource versus for multiple resources.

By consistently encapsulating the requested resources in an array, regardless of whether one or many resources were requested, there will only need to be a single consistent client-side process for parsing the response data. This will make the client-side code more clear and concise.