Specify a fix to a small security vulnerability
It appears that there is a security vulnerability that stems from using an array as the top-level object in an HTTP response. See https://haacked.com/archive/2008/11/20/anatomy-of-a-subtle-json-vulnerability.aspx/

This commit specifies changes that would address this vulnerability by encapsulating the response array under the key "messages" within a top-level JSON object.