Implement security fix
This commit implements changes that address a security vulnerability that partially stems from using an array as the top-level object in an HTTP response. See https://haacked.com/archive/2008/11/20/anatomy-of-a-subtle-json-vulnerability.aspx/

It looks like this vulnerability would not apply under the current architecture, since it also relies on having users who are authenticated in a session that the browser remembers. This application's architecture doesn't work that way (it authenticates via an HMAC in each HTTP request that requires the correct key, and has nothing to do with an authentication session). But... just in case I forget about this later and change the app to an architecture with logged-in users, I'm still going to apply these changes.

To fix this possible vulnerability, these changes encapsulate the response array under the key "messages" within a top-level JSON object. From now on, this vulnerability should no longer be an issue, regardless of whether I keep the same authentication scheme or change it.