Add a unit spec for generating a digital signature
This commit specifies the method Authentication#generate_signature(message). This method should take a string and return a SHA-256 HMAC of that string.

Eventually, this will be used to sign an HTTP response. This will enable the client to authenticate the sender of the response and to verify that the response was not tampered with during transit.