Specify signatures in the HTTP response
Each bottom-level object in the response body should include a digital signature. The signature is a SHA-256 HMAC hash of the original unencrypted message data.

This will enable the client to verify the origin and authenticity of the response by generating a signature using the same methods as the server, and comparing that signature to the signature in the response.

Also, this commit removes a few duplicate specifications. There were a few specifications that the response should have certain keys, which were followed by specifications of the types of the values held in those keys. The latter specifications encompass the behavior of the former specifications, so I removed the former specifications. This slighly simplifies the tests while still encompassing all of the desired behavior.